hitconctf-2024-final-writeup/ad-fulu/decompiler/requirement.txt
pillow
aggdraw